requirements.txt
matplotlib==3.9.2
numpy==1.26.2
pandas==2.0.3
pydantic==2.0.3
shapely==2.0.6
skyfield==1.48
cartopy==0.22.0
geopandas==1.0.1
pillow==10.0.0
PyYAML==6.0.1
pyarrow==17.0.0
pyogrio==0.10.0
rtree==1.3.0
requests==2.31.0
duckdb==1.1.3
ibis-framework[duckdb,geospatial]==9.5.0
